requirements.txt
peewee==3.9.3
python-dotenv==0.10.1
python-telegram-bot==11.1.0